+++ Automatic_Inventory_Manager/Mods/Automatic_Inventory_Manager/ScriptExtender/Lua/AIM/FilterPresets/_AllDefaults.lua
@@ -2,9 +2,8 @@ local all_defaults = {}
 all_defaults.Weapons = {
 	[ItemFilters.ItemKeys.WILDCARD] = {
 		Filters = {
-			[90] = { TargetStat = ItemFilters.FilterFields.TargetStat.HAS_TYPE_EQUIPPED },
-			[91] = { TargetStat = ItemFilters.FilterFields.TargetStat.WEAPON_ABILITY, CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER },
-			[92] = { TargetStat = ItemFilters.FilterFields.TargetStat.WEAPON_SCORE, CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER },
+			[90] = { TargetStat = ItemFilters.FilterFields.TargetStat.WEAPON_ABILITY, CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER },
+			[91] = { TargetStat = ItemFilters.FilterFields.TargetStat.WEAPON_SCORE, CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER },
 		}
 	}
 }
@@ -12,11 +11,20 @@ all_defaults.Weapons = {
 all_defaults.Equipment = {
 	[ItemFilters.ItemKeys.WILDCARD] = {
 		Filters = {
-			[50] = {
+			[50] = { TargetStat = ItemFilters.FilterFields.TargetStat.HAS_TYPE_EQUIPPED },
+			[51] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
 			},
-			[51] = { TargetStat = ItemFilters.FilterFields.TargetStat.PROFICIENCY },
+			[52] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["EQUIPMENT_TYPES"] = { "SELF" },
+					["ARMOR_TYPES"] = { "SELF" }
+				},
+			},
+			[53] = { TargetStat = ItemFilters.FilterFields.TargetStat.PROFICIENCY },
 		}
 	}
 }
@@ -24,16 +32,16 @@ all_defaults.Equipment = {
 all_defaults.Roots = {
 	-- not a typo :D
 	["ALCH_Soultion_Elixir_Barkskin_cc1a8802-675a-426b-a791-ec1d5a5b6328"] = {
-		PreFilters = { [ItemFilters.ItemFields.PreFilters.STACK_LIMIT] = 1 },
+		PreFilters = { [ItemFilters.ItemFields.PreFilters.STACK_LIMIT] = 2 },
 		Filters = {
-			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.ARMOR_CLASS, CompareStategy = ItemFilters.FilterFields.CompareStategy.LOWER }
+			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.ARMOR_CLASS, CompareStrategy = ItemFilters.FilterFields.CompareStrategy.LOWER }
 		}
 	},
 	["LOOT_Gold_A_1c3c9c74-34a1-4685-989e-410dc080be6f"] = {
 		Filters = {
 			[1] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			}
 		}
 	}
@@ -43,25 +51,34 @@ all_defaults.Tags = {
 	["HEALING_POTION"] = {
 		PreFilters = { [ItemFilters.ItemFields.PreFilters.STACK_LIMIT] = 2 },
 		Filters = {
-			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.HEALTH_PERCENTAGE, CompareStategy = ItemFilters.FilterFields.CompareStategy.LOWER, },
-			[2] = { TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT, CompareStategy = ItemFilters.FilterFields.CompareStategy.LOWER }
+			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.HEALTH_PERCENTAGE, CompareStrategy = ItemFilters.FilterFields.CompareStrategy.LOWER, },
+			[2] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.LOWER,
+				CalculateStackUsing = {
+					["TAGS"] = { "HEALING_POTION" }
+				},
+			}
 		},
 	},
 	["LOCKPICKS"] = {
 		Filters = {
-			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.SKILL_TYPE, TargetSubStat = "SleightOfHand", CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER, },
+			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.SKILL_TYPE, TargetSubStat = "SleightOfHand", CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER, },
 			[2] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			}
 		},
 	},
 	["TOOL"] = {
 		Filters = {
-			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.SKILL_TYPE, TargetSubStat = "SleightOfHand", CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER, },
+			[1] = { TargetStat = ItemFilters.FilterFields.TargetStat.SKILL_TYPE, TargetSubStat = "SleightOfHand", CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER, },
 			[2] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["TAGS"] = { "TOOL" }
+				}
 			}
 		},
 	},
@@ -69,43 +86,71 @@ all_defaults.Tags = {
 		Filters = {
 			[1] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			},
-			[2] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Dexterity", CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER }
+			[2] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["TAGS"] = { "COATING" }
+				},
+			},
+			[3] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Dexterity", CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER }
 		}
 	},
 	["ARROW"] = {
 		Filters = {
 			[1] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			},
-			[2] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Dexterity", CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER }
+			[2] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["TAGS"] = { "ARROW" }
+				},
+			},
+			[3] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Dexterity", CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER }
 		}
 	},
 	["GRENADE"] = {
 		Filters = {
 			[1] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			},
-			[2] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Strength", CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER }
+			[2] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["TAGS"] = { "GRENADE" }
+				},
+			},
+			[3] = { TargetStat = ItemFilters.FilterFields.TargetStat.ABILITY_STAT, TargetSubStat = "Strength", CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER }
 		}
 	},
 	["SCROLL"] = {
 		Filters = {
 			[1] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
 			},
-			[2] = { Target = "originalTarget" },
+			[2] = {
+				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER,
+				CalculateStackUsing = {
+					["TAGS"] = { "SCROLL" }
+				},
+			},
+			[3] = { Target = "originalTarget" },
 		}
 	},
 	["CONSUMABLE"] = {
 		Filters = {
 			[99] = {
 				TargetStat = ItemFilters.FilterFields.TargetStat.STACK_AMOUNT,
-				CompareStategy = ItemFilters.FilterFields.CompareStategy.HIGHER
+				CompareStrategy = ItemFilters.FilterFields.CompareStrategy.HIGHER
 			}
 		},
 	},