+++ Automatic_Inventory_Manager/Mods/Automatic_Inventory_Manager/ScriptExtender/Lua/AIM/_Config.lua
@@ -28,13 +28,10 @@ local function InitializeConfigurations()
 	FileUtils:SaveTableToFile("config.json", Config.AIM)
 end
 
---- Initializes the presets from AIM/FilterPresets/, overwriting directory contents, and migrates 1.x FILTER_DIR and FILTER_TABLES config properties to new structure,
---- copying existing files over to a custom preset directory and activating it. If no legacy configs are detected, sets Preset_AllDefaults as the active preset
-local function InitializeFilterPresetsAndUpgradeLegacyFilters()
+--- Initializes the presets from AIM/FilterPresets/, overwriting directory contents
+local function InitializeFilterPresets()
 	ItemFilters:RegisterItemFilterMapPreset(ModUtils:GetAIMModInfo().ModuleUUID, Preset_AllDefaults.Name, Preset_AllDefaults.ItemFilterMaps)
 	ItemFilters:RegisterItemFilterMapPreset(ModUtils:GetAIMModInfo().ModuleUUID, Preset_CampGoldBooks.Name, Preset_CampGoldBooks.ItemFilterMaps)
-
-	Upgrade:LegacyFiltersToPresets()
 
 	if not Config.AIM.PRESETS.ACTIVE_PRESETS or next(Config.AIM.PRESETS.ACTIVE_PRESETS) == nil then
 		Config.AIM.PRESETS.ACTIVE_PRESETS = { [ModUtils:GetAIMModInfo().Name .. "-" .. Preset_AllDefaults.Name] = { "ALL" } }
@@ -67,11 +64,9 @@ function Config.SyncConfigsAndFilters()
 		config.PRESETS.FILTERS_PRESETS[presetName] = preset
 	end
 
-	Upgrade:ConfigFile(config)
-
 	Config.AIM = config
 
-	InitializeFilterPresetsAndUpgradeLegacyFilters()
+	InitializeFilterPresets()
 
 	ItemFilters:LoadItemFilterPresets()
 
@@ -79,6 +74,6 @@ function Config.SyncConfigsAndFilters()
 
 	EntityPropertyRecorder:LoadRecordedItems()
 
-	Logger:BasicInfo("AIM has finished initialization in %dms!", Ext.Utils.MonotonicTime() - startTime)
 	Config.IsInitialized = true
+	Logger:BasicInfo("AIM has finished initialization in %dms!", Ext.Utils.MonotonicTime() - startTime)
 end